move news controller

+++ app/Http/routes.php
@@ -36,13 +36,13 @@ Route::get('home', 'HomeController@index');
 Route::group(['prefix' => 'admin', 'middleware' => 'auth','admin'], function (){
     Route::get('/',['as' => 'admin','uses' => 'Admin\AdminController@index']);
     
-    Route::get('/news', ['as' => 'news.index', 'uses' => 'NewsController@index']);
-    Route::get('/news/create', ['as' => 'news.create', 'uses' => 'NewsController@create']);
-    Route::post('/news', ['as' => 'news.store', 'uses' => 'NewsController@store']);
-    Route::get('/news/{id}', ['as' => 'news.show', 'uses' => 'NewsController@show']);
-    Route::get('/news/{id}/edit', ['as' => 'news.edit', 'uses' => 'NewsController@edit']);
-    Route::put('/news/{id}', ['as' => 'news.update', 'uses' => 'NewsController@update']);
-    Route::delete('/news/{id}', ['as' => 'news.destroy', 'uses' => 'NewsController@destroy']);
+    Route::get('/news', ['as' => 'news.index', 'uses' => 'Admin\NewsController@index']);
+    Route::get('/news/create', ['as' => 'news.create', 'uses' => 'Admin\NewsController@create']);
+    Route::post('/news', ['as' => 'news.store', 'uses' => 'Admin\NewsController@store']);
+    Route::get('/news/{id}', ['as' => 'news.show', 'uses' => 'Admin\NewsController@show']);
+    Route::get('/news/{id}/edit', ['as' => 'news.edit', 'uses' => 'Admin\NewsController@edit']);
+    Route::put('/news/{id}', ['as' => 'news.update', 'uses' => 'Admin\NewsController@update']);
+    Route::delete('/news/{id}', ['as' => 'news.destroy', 'uses' => 'Admin\NewsController@destroy']);
 });
 
 //------------SHOW NEWS------------------------------------------------------
